completed the header section
