+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-import { Header } from "@/sections/header";
+import { Header } from "@/sections/TheHeader";
+
 
 export default function Home() {
   return (
-    <>
-      <Header />
+    <><Header />
     </>
   );
 }
