+++ src/sections/TheHeader.tsx
@@ -1,0 +1,35 @@
+import ArrowRight from "@/icons/ArrowRight";
+import LogoIcon from "@/assets/logosaas.png";
+import MenuIcon from "@/icons/Menu";
+import Image from "next/image";
+
+export const Header = () => {
+  return (
+    <header className="sticky top-0">
+      <div className="flex justify-center items-center py-3 bg-black text-white text-sm gap-3">
+        <p className="hidden md:block text-white/60">Streamline your workflow and boost your productivity</p>
+        <div className="inline-flex gap-1 items-center">
+          <p>Get started for free</p>
+          <ArrowRight className="h-4 w-4 inline-flex justify-center items-center" />
+        </div>
+      </div>
+      <div className="py-5">
+        <div className="container">
+          <div className="flex items-center justify-between">
+            <Image src={LogoIcon} alt="Saas Logo" height={40} width={40} />
+            <MenuIcon className="h-5 w-5 md:hidden" />
+            <nav className="hidden md:flex gap-6 text-black/60 items-center">
+                
+                  <a href="#">About</a>
+                  <a href="#">About</a>
+                  <a href="#">About</a>
+                  <a href="#">About</a>
+                  <a href="#">About</a>
+                <button className="bg-black text-white px-4 py-2 rounded-lg font-medium inline-flex items-center justify-center tracking-tight">Get for free</button>
+            </nav>
+          </div>
+        </div>
+      </div>
+    </header>
+  );
+};
