--- src/sections/header.tsx
@@ -1,24 +1,0 @@
-import ArrowRight from "@/icons/ArrowRight";
-import LogoIcon from "@/assets/logosaas.png";
-import MenuIcon from "@/assets/menu.svg";
-import Image from "next/image";
-
-export const Header = () => {
-  return (
-    <header>
-      <div className="flex justify-center items-center py-3 bg-black text-white text-sm">
-        <div className="inline-flex gap-1 items-center">
-          <p>Get started for free</p>
-          <ArrowRight className="h-4 w-4 inline-flex justify-center items-center" />
-        </div>
-      </div>
-      <div className="py-5">
-        <div className="container">
-          <Image src={LogoIcon} alt="Saas Logo" height={40} width={40} />
-          {/* <Image src={MenuIcon} alt="Menu Logo" height={40} width={40} />
-          <MenuIcon /> */}
-        </div>
-      </div>
-    </header>
-  );
-};
